refactor: consolidate intro and skills sections into main page
